AddNoise: Rename to AddGaussianNoise
This op was renamed to make it clear that the noise generated by this op
is NON-uniform. This distinction is being made to separate this op from
a uniform noise generation op (coming in a later commit)
