+++ src/main/java/net/imagej/ops/filter/FilterNamespace.java
@@ -67,36 +67,36 @@ public class FilterNamespace extends AbstractNamespace {
 
 	// -- addNoise --
 
-	@OpMethod(ops = { net.imagej.ops.filter.addNoise.AddNoiseRealType.class,
-		net.imagej.ops.filter.addNoise.AddNoiseRealTypeCFI.class })
-	public <I extends RealType<I>, O extends RealType<O>> O addNoise(final O out,
+	@OpMethod(ops = { net.imagej.ops.filter.addGaussianNoise.AddGaussianNoiseRealType.class,
+		net.imagej.ops.filter.addGaussianNoise.AddGaussianNoiseRealTypeCFI.class })
+	public <I extends RealType<I>, O extends RealType<O>> O addGaussianNoise(final O out,
 		final I in, final double rangeMin, final double rangeMax,
 		final double rangeStdDev)
 	{
 		@SuppressWarnings("unchecked")
-		final O result = (O) ops().run(Ops.Filter.AddNoise.class, out, in, rangeMin,
+		final O result = (O) ops().run(Ops.Filter.AddGaussianNoise.class, out, in, rangeMin,
 			rangeMax, rangeStdDev);
 		return result;
 	}
 
-	@OpMethod(ops = { net.imagej.ops.filter.addNoise.AddNoiseRealType.class,
-		net.imagej.ops.filter.addNoise.AddNoiseRealTypeCFI.class })
-	public <I extends RealType<I>, O extends RealType<O>> O addNoise(final O out,
+	@OpMethod(ops = { net.imagej.ops.filter.addGaussianNoise.AddGaussianNoiseRealType.class,
+		net.imagej.ops.filter.addGaussianNoise.AddGaussianNoiseRealTypeCFI.class })
+	public <I extends RealType<I>, O extends RealType<O>> O addGaussianNoise(final O out,
 		final I in, final double rangeMin, final double rangeMax,
 		final double rangeStdDev, final long seed)
 	{
 		@SuppressWarnings("unchecked")
-		final O result = (O) ops().run(Ops.Filter.AddNoise.class, out, in, rangeMin,
+		final O result = (O) ops().run(Ops.Filter.AddGaussianNoise.class, out, in, rangeMin,
 			rangeMax, rangeStdDev, seed);
 		return result;
 	}
 
-	@OpMethod(op = net.imagej.ops.filter.addNoise.AddNoiseRealTypeCFI.class)
-	public <T extends RealType<T>> T addNoise(final T in, final double rangeMin,
+	@OpMethod(op = net.imagej.ops.filter.addGaussianNoise.AddGaussianNoiseRealTypeCFI.class)
+	public <T extends RealType<T>> T addGaussianNoise(final T in, final double rangeMin,
 		final double rangeMax, final double rangeStdDev)
 	{
 		@SuppressWarnings("unchecked")
-		final T result = (T) ops().run(Ops.Filter.AddNoise.class, in, rangeMin,
+		final T result = (T) ops().run(Ops.Filter.AddGaussianNoise.class, in, rangeMin,
 			rangeMax, rangeStdDev);
 		return result;
 	}
