+++ src/main/templates/net/imagej/ops/Ops.list
@@ -96,7 +96,7 @@ namespaces = ```
 		
 	]],
 	[name: "filter", iface: "Filter", ops: [
-		[name: "addNoise",                       iface: "AddNoise"],
+		[name: "addGaussianNoise",               iface: "AddGaussianNoise"],
 		[name: "addPoissonNoise",                iface: "AddPoissonNoise"],
 		[name: "bilateral",                      iface: "Bilateral"],
 		[name: "convolve",                       iface: "Convolve"],
