+++ src/test/java/net/imagej/ops/convert/ConvertIIsTest.java
@@ -143,7 +143,7 @@ public class ConvertIIsTest extends AbstractOpTest {
 	private void addNoise(final Iterable<ShortType> image) {
 		// NB: Need "? super ShortType" instead of "?" to make javac happy.
 		final UnaryInplaceOp<? super ShortType, ShortType> noiseOp = Inplaces.unary(
-			ops, Ops.Filter.AddNoise.class, ShortType.class, -32768, 32767, 10000);
+			ops, Ops.Filter.AddGaussianNoise.class, ShortType.class, -32768, 32767, 10000);
 		ops.map(image, noiseOp);
 	}
